# Optimize formatTime utility