Pull request checklist:

- [x] `CHANGELOG.md` was updated, if applicable

# Description

Cherry-picked from #3775

<details>
<summary>Benchmark results</summary>
built on commit a5f081268e891dc7ea62db3b15e2d98296275711 with `cmake -DBUILD_BENCHMARKS=On -DBUILD_APP=Off -DUSE_PRECOMPILED_HEADERS=Off -DCMAKE_BUILD_TYPE=Debug`

```
2022-05-29T00:20:03+02:00
Running ./bin/chatterino-benchmark
Run on (16 X 1983.15 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x8)
  L1 Instruction 32 KiB (x8)
  L2 Unified 512 KiB (x8)
  L3 Unified 4096 KiB (x2)
Load Average: 1.33, 1.63, 1.82
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
BM_ShortcodeParsing        4250 ns         4249 ns       164307
BM_TimeFormatting          2322 ns         2321 ns       301147
BM_TimeFormattingOld       4169 ns         4168 ns       167841
```
</details>